+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 ### Added
 - Added XPath evaluation functions with runtime tpye checking of the results of the expressions [[#181]](https://github.com/JuDFTteam/masci-tools/pull/181)
 - Command `masci-tools fleur-schema pull <branch>` to update/add Fleur Schema files from the iffgit more easily [[#184]](https://github.com/JuDFTteam/masci-tools/pull/184)
+- New XML setters [[#183]](https://github.com/JuDFTteam/masci-tools/pull/183):
+  - Setting XC functional explicitly + LibXC support (`set_xcfunctional`),
+  - Creating a kpoint path using `ase` (`set_kpointpath`)
+  - Creating a kpoint mesh with symmetry reduction using `spglib`. Should be equivalent to the `gamma@grid=nx,ny,nz` kpoint generator in `inpgen` (`set_kpointmesh`)
 
 ### Bugfixes
 - Add clearer error message if `None` is passed to the `convert_to_xml` functions. This would happen for example using the `set_inpchanges` function with `{'minDistance': None}` [[#182]](https://github.com/JuDFTteam/masci-tools/pull/182)